requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.39.1
django-cors-headers==3.1.1
django-filter==2.1.0
djangorestframework==3.9.4
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
Markdown==3.1
more-itertools==7.2.0
oauthlib==3.0.1
psycopg2-binary==2.8.2
python3-openid==3.1.0
pytz==2019.1
redis==3.3.8
requests==2.21.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.24.3
vine==1.3.0
zipp==0.6.0
